Automatically start a NodeJS Express server when running the application
The node express server currently performs two tasks:
- It serves the production vite code
- It serves as a CORS proxy for two hosts

The server will shut down automatically when the application terminates.

Note that the server will fail to start in development mode,
because vite already has the port open. This is OK because the
vite server already serves the development code and CORS proxy.