AutoConfig: For Office365, also check MX long domain
MX for Office365 domains is e.g. eu.mail.protection.outlook.com ,
so MX baseDomain is office.com, but office.com is private accounts
and a different config. (Thanks Microsoft, for confusing names all the time.)
So, per spec, we need to look up mail.protection.outlook.com in the ISPDB,
which gives the Office365 config.