EWS, OWA, ActS: Fix import of ResponseType (#342)
`OWAEvent.ts` was using `import { type ResponseType }`, although its values need to be accessible. This doesn't cause a run-time error because the global `ResponseType` object gets used instead, so instead the code just fails silently.

Additionally the import filename was overly long (caused by a move of the file) and this trims the import and groups it with the other imports from the same directory.

The overly long import filename also happens in two other files but in this case that's the only change as they weren't using `import { type ResponseType }`.