LiveKit: Show video when participant joins
We first get an audio track, then a video track.
We add a `VideoStream` for the audio track, then set `hasVideo = true` when
the video is added.
However, the UI observers use `videos.filter(v => v.hasVideo)`, and they
don't observe the target objects, so they don't know about the property
change.

This is a known limitation of `Collection.filter()`. Attempts to fix it
in svelte-collections 1.2.6 work in unit tests, but not in the app.
See branch `ben/filter-observable`. This needs to be debugged further.

This works around the problem by manually refreshing the lists.