Calendar: Update UI after simple event is deleted or made recurring (#687)
`mapToCollection` has a problem, and that is that it assumes that the mapping function is idempotent, i.e. it always returns the same value each time.

This is not true with our current `mapToEvents` function, except in the edge case where you only create and delete recurring events (which does work on current code).

Now there may be another way to achieve this which allows a dynamic mapping function, but I chose to go the other way and make the mapping function idempotent.

To this end, the `instances` collection of the event should contain itself if it's a single event or an exception otherwise it should contain its occurrences.

While I was there I noticed that changing a recurring event to a single event didn't remove its occurrences (because clearExceptions regenerates them) so I rearranged the code so that this works too.