UI: Person Autocomplete: Allow to create Participant objects
This fixes a bug "introduced" by eaf6fe3.

This PR now uses the approach from PR #885 of switching from an event to a
callback. The default callback simply adds or removes the person from the
collection, but Calendar can override the `onAddPerson()` to add a
`Participant` instead, thus preserving type safety.
Meanwhile PR #885 also wants to be able to override the `onRemovePerson()`
callback as removing delegates happens immediately (as opposed to removing
permissions which happens once editing of the permissions is complete).