EWS, ActS: Fast fail when OAuth2 isn't logged in (#967)
When an EWS request is made, there's no actual check that the OAuth2 object is logged in. By comparison with OWA, the request fails immediately if it hasn't been logged in.

This is one way to fix a race condition that exists when you try to log in to EWS via the in-app tab. (The race condition causes the OAuth2 object to log out when the racing request fails, even though it is logging in at the time.)

* EWS: Fast fail when OAuth2 isn't logged in
* ActiveSync: Fast fail when OAuth2 isn't logged in